test: replacing each `DocumentStore` specific tests and used the generalised ones from `haystack.testing.document_store` (#2812)
* removing test_delete_all_documents from PgVector - already in haystack.testing

* removing duplicated tests from OpenSearchDocumentStore

* removing duplicated tests from ElasticStoreDocumentTest

* removing test_delete_all_documents from weaviate

* removing test_delete_all_documents from qdrant

* removing delete_all_documents tests from PineCone

* removing delete_all_documents tests from MondoDBDocumentStore

* adding check for empty list for delete_all_documents on Chroma

* removing delete_all_documents tests from AzureAISearchDocumentStore

* fixing Pinecone tests

* cleaning up tests in qdrant and weaviate

* removing more weavite delete_by_filter tests

* cleaning up tests in pgvector

* updating OpenSearch: delete_by_filter_async refresh=True as default

* updating OpenSearch: update_by_filter refresh=True as default

* removing standard test from MongoDBDocumentStore

* reverting default refresh param in OpenSearch

* removing standard test from Chroma

* updating AzuresAISearchDocumentStore to make use of base tests

* removing duplicated tests from Astra

* adding UpdateByFilterTest to Weavite

* fixing Azure linting issues

* formatting

* formatting

* temporarly using haystack branch with new standard tests

* temporarly using haystack branch with new standard tests

* formatting

* removing duplicated tests from Pinecone

* updating all tests to use new test class from haystack core

* fixing imports for 3.10

* fixing linting issues

* fixing azure_ai_search overrides due to class changes in haystack core

* fixing linting

* chaning imports due to MRO errors + changing document_stores fixtures when needed due to new method from ExtendedTests

* fixing weaviate tests and imports

* fixing Qdrandt tests and imports

* fixing PGVector tests

* fixing Chroma tests + imports

* pointing all pyproject.toml to main

* fixing elasticsearch 3.10 arg not used

* fixing mongodb imports

* fixing chroma formatting

* fixing weaviate tests and imports

* fixing pinecone imports

* attending PR comments

* removing white space

* adding filterable docs fixture to azure_ai_search

* fixing linter

* updating pyproject.toml to reference pypi instead of main branch

* updating pyproject.toml to reference latest pypi instead of main branch

* Update integrations/qdrant/pyproject.toml


* Update integrations/pgvector/pyproject.toml


* Update integrations/mongodb_atlas/pyproject.toml


* Update integrations/weaviate/pyproject.toml


---------

+++ integrations/chroma/src/haystack_integrations/document_stores/chroma/document_store.py
@@ -856,12 +856,13 @@ class ChromaDocumentStore:
             else:
                 collection = self._collection.get()
                 ids = collection.get("ids", [])
-                self._collection.delete(ids=ids)  # type: ignore
-                logger.info(
-                    "Deleted all the {n_docs} documents from the collection '{name}'.",
-                    name=self._collection_name,
-                    n_docs=len(ids),
-                )
+                if ids:
+                    self._collection.delete(ids=ids)  # type: ignore
+                    logger.info(
+                        "Deleted all the {n_docs} documents from the collection '{name}'.",
+                        name=self._collection_name,
+                        n_docs=len(ids),
+                    )
         except Exception as e:
             msg = f"Failed to delete all documents from ChromaDB: {e!s}"
             raise DocumentStoreError(msg) from e
@@ -895,12 +896,13 @@ class ChromaDocumentStore:
             else:
                 collection = await self._async_collection.get()
                 ids = collection.get("ids", [])
-                await self._async_collection.delete(ids=ids)  # type: ignore
-                logger.info(
-                    "Deleted all the {n_docs} documents from the collection '{name}'.",
-                    name=self._collection_name,
-                    n_docs=len(ids),
-                )
+                if ids:
+                    await self._async_collection.delete(ids=ids)  # type: ignore
+                    logger.info(
+                        "Deleted all the {n_docs} documents from the collection '{name}'.",
+                        name=self._collection_name,
+                        n_docs=len(ids),
+                    )
 
         except Exception as e:
             msg = f"Failed to delete all documents from ChromaDB: {e!s}"
